feat(frontend): move auth buttons to header
